+++ src/controllers/getCard.js
@@ -1,8 +1,8 @@
-import requestInstance from "../requestInstance/requestInstance";
+import httpClient from "../httpClient/httpClient";
 
 const getCard = async () => {
   const randNum = Math.ceil(Math.random() * 100);
-  const { data: card } = await requestInstance.get(`/posts/${randNum}`);
+  const { data: card } = await httpClient.get(`/posts/${randNum}`);
   return card;
 };
